refactor(cloudformation): split template.rs by concerns (3826 -> 2238 mod + 6 submodules)
Audit 2026-05-19 P8 remainder. template.rs split into concern submodules:

- intrinsics.rs   629  (resolve_refs / pseudo_value / no_value sentinel /
                       stringify_value / compute_cidr_subnets / parse_getatt)
- for_each.rs     366  (expand_for_each / expand_sam / resolve_for_each_items /
                       substitute_loop_vars{,in_value})
- parser.rs       277  (parse_template* + collect_import_value_names +
                       parse_outputs + import walkers)
- mappings.rs     190  (parse_mappings / apply_mappings / resolve_find_in_map /
                       stringify_findinmap_arg)
- conditions.rs   119  (evaluate_conditions / evaluate_condition_named /
                       eval_condition_expr)
- resolution.rs    77  (resolve_resource_properties{,_with_attrs})

mod.rs retains struct defs (ParsedTemplate, TemplateOutput, ResourceDefinition),
the test corpus, and pub re-exports for the externally-visible parser/resolution
fns so callers like resource_provisioner keep their existing paths.

201 unit tests pass.
